This commit is really just a checkpoint...
we currenlty have, for pgs, add, delete, cut, copy and paste.
We are experimenting with a "timeout" method for deactivating
cells in the ink and page ring. And we're looking at improvments
to the Symbols panel (which should be cazlled Symbol, not Symbols,
fwiw). but we now want to get have page's add/delete lbe strictly
local, with their own undo stack. And the pastepanel will show
thumbnails, not line numbers, because line numbers are really
transitent.